+++ src/doomstat.c
@@ -164,6 +164,7 @@ int menu_background_all;
 int no_menu_tint;
 int no_berserk_tint;
 int no_radsuit_tint;
+int nightvision_visor;
 int damagecount_cap;
 int bonuscount_cap;
 int fake_contrast;