+++ src/p_user.c
@@ -911,7 +911,9 @@ void P_PlayerThink (player_t* player)
 
     player->powers[pw_invulnerability] > 4*32 ||    /* Regular Doom */
     player->powers[pw_invulnerability] & 8 ? INVERSECOLORMAP :
-    player->powers[pw_infrared] > 4*32 || player->powers[pw_infrared] & 8;
+    player->powers[pw_infrared] > 4*32 || player->powers[pw_infrared] & 8
+    // [Nugget] Night-vision visor
+    ? (STRICTMODE(nightvision_visor) ? 33 : 1) : 0;
 }
 
 //----------------------------------------------------------------------------